Classes organizadas em seus arquivos-fonte, e armazenadas na nova pasta Classes.
